general refactore, got rid of context. Lifted state to the App component
