+++ src/app/App.tsx
@@ -1,32 +1,39 @@
 import * as React from 'react'
-import { Reset } from './styles/global'
-import {
-  BoardManagerContext,
-  FrameManagerContext,
-  SnakeManagerContext,
-} from './managers/context'
+import { hot } from 'react-hot-loader'
 import BoardManager from './managers/BoardManager'
-import Board from './components/Board'
 import FrameManager from './managers/FrameManager'
 import SnakeManager from './managers/SnakeManager'
+import { Position } from './managers/Snake/SnakeNode'
+import { addEventListeners } from './utils'
+import Board from './components/Board'
+
+import { Reset } from './styles'
 
 const boardManager = new BoardManager()
 const frameManager = new FrameManager()
 const snakeManager = new SnakeManager()
 
 const App: React.FC = () => {
+  const [snake, setSnake] = React.useState<Position[]>([])
+
+  const runFrameUpdate = React.useCallback(() => {
+    snakeManager.moveSnake()
+    const snake = snakeManager.getSnake()
+    setSnake(snake)
+  }, [])
+
+  React.useEffect(() => {
+    setSnake(snakeManager.getSnake())
+    frameManager.init(runFrameUpdate)
+    addEventListeners(snakeManager)
+  }, [])
+
   return (
     <>
-      <FrameManagerContext.Provider value={frameManager}>
-        <SnakeManagerContext.Provider value={snakeManager}>
-          <BoardManagerContext.Provider value={boardManager}>
-            <Reset />
-            <Board />
-          </BoardManagerContext.Provider>
-        </SnakeManagerContext.Provider>
-      </FrameManagerContext.Provider>
+      <Reset />
+      <Board snake={snake} cells={boardManager.getBoard()} />
     </>
   )
 }
 
-export default App
+export default hot(module)(App)
