--- src/app/components/hooks.ts
@@ -1,6 +1,0 @@
-import { useFrameManager } from '../managers/context'
-
-export const useStartGame = (callBack: () => void) => {
-  const frameManager = useFrameManager()
-  frameManager.init(callBack)
-}
